Moved polylog, zeta, and related functions into SpecialFunctions.
